Add yohann's work
If you want to try it before work on it, you must create a database with
the file SQLScript.sql in the folder database, and change the connection
values in the file config.php (values which correspond to your server
configuraton) in the config file.

Also you have to start the URL Rewriting on your local server (a basic
request on google can show you how to do that) if it doesn't work, tell
me and I can explain you in detail.

Also, you can check my work on yohann-haution.fr/HW-project

The login/password is admin/admin