Added umpire management
Pretty much the same code and layout of the tournament management.  I
didn't think it was necessary to show the availability in the table but
this can be changed if required.

When adding umpires, the selected checkboxes are remembered for when you
add the next.  I still need to fix that.